move routes to separate files and refactor modules
