+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nicklinref"
-version = "0.10.1"
+version = "0.10.2"
 authors = ["thehappycheese"]
 edition = "2021"
 readme = "README.md"
