+++ changelog.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## [0.10.0] 2023-06-01 - Unreleased
+## [0.10.2] 2023-06-05 - Unreleased
+
+- Refactoring to make main.rs short and easy to understand
+
+## [0.10.1] 2023-06-01
 
 - create changelog
 - enable gzip compression
 - implement `x-request-id` echo header
+  - note: still does not work on reject :/
