+++ src/data/esri_serde.rs
@@ -26,10 +26,11 @@ impl Ord for LayerFeatureAttr {
 			std::cmp::Ordering::Less=>std::cmp::Ordering::Less,
 			std::cmp::Ordering::Equal=>{
 				
-				// The current implementation does not rely on ordering by START_SLK. We can finish the algorithm earlier:
-				self.CWY.cmp(&other.CWY)
+                self.CWY.cmp(&other.CWY)
+				// TODO: The current implementation does not rely on ordering 
+                // by START_SLK, therefore we can stop here before sorting.
+				// don't forget the sort was removed if you re-write the lookup.
 
-				// TODO: don't forget you removed this:
 				// match self.CWY.cmp(&other.CWY){
 				// 	std::cmp::Ordering::Less=>std::cmp::Ordering::Less,
 				// 	std::cmp::Ordering::Equal=>{
