+++ src/data/mod.rs
@@ -1,0 +1,9 @@
+mod download_or_read_from_file;
+pub use download_or_read_from_file::{
+    load_data_from_file,
+    update_data_from_service
+};
+
+pub mod index;
+
+pub mod esri_serde;
