+++ src/helpers/echo_x_request_id.rs
@@ -2,8 +2,10 @@ use warp::{http::HeaderValue, reply::Response, Filter, Rejection, Reply};
 
 /// Echoes back the `x-request-id` header from the request, if it is present and
 /// can be parsed as a `u64`. Otherwise the response is not modified.
-/// If the header is present and valid it will be echoed even if the response
-/// from `filter` is a rejection
+/// 
+/// TODO: the desired behavior is that if the header is present and valid it
+/// will be echoed even if the response from `filter` is a rejection.
+/// Currently I don't see a way to accomplish that...
 ///
 /// # Example
 ///
