+++ src/routes/geoprocessing/get_points.rs
@@ -1,6 +1,6 @@
-use crate::esri_serde::{LayerSaved};
-use crate::query_parameters::{OutputFormat, QueryParametersPoint};
-use crate::update_data::{LookupMap, RoadDataByCwy};
+use crate::data::esri_serde::{LayerSaved};
+use crate::routes::query_parameters::{OutputFormat, QueryParametersPoint};
+use crate::data::index::{LookupMap, RoadDataByCwy};
 use nickslinetoolsrust::line_string_measured::LineStringMeasured;
 use nickslinetoolsrust::vector2::Vector2;
 use std::str;
