+++ src/routes/mod.rs
@@ -1,0 +1,12 @@
+mod geoprocessing;
+
+pub mod query_parameters;
+
+mod lines;
+pub use lines::lines;
+
+mod points;
+pub use points::points;
+
+mod lines_batch;
+pub use lines_batch::lines_batch;
