+++ src/routes/query_parameters/query_parameters_line.rs
@@ -39,7 +39,7 @@ fn default_output_format() -> OutputFormat {
 }
 
 fn default_m()->bool{
- 	false
+	false
 }
